Refactor user-related services and controllers; move AppUserDetailsService to util package and implement CustomerService for customer management
